Update February parameters, remove month_idx
It occurred that version with 0.4 percentage distribution for February was uploaded to the competition, so replace 0.45 DISTR_PERC with 0.4.

Removed month_idx from the training pipeline as it is unnecessary and result in a bug when not all months are trained at once. Now it takes[month - 1] which should be always the correct month